requirements.txt
Flask = 2.2.5
Werkzeug = 2.2.3
Pillow = 9.5.0
opencv-python = 4.9.0.72
